pyproject.toml
[project]
name = "golangci-lint-linter"
dynamic = ["version"]
description = "Linter for the golanci-lint files."
readme = "README.md"
authors = [
    { name = "Manuel Doncel Martos", email = "manueldoncelmartos@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.2.1",
    "ruamel-yaml>=0.18.14",
]
license = "MIT AND (Apache-2.0 OR BSD-2-Clause)"
keywords = ["linter", "golangci-lint", "go", "best-practices"]
classifiers = [
  "Development Status :: 3 - Alpha",

  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",

  "Programming Language :: Python",
]

[project.scripts]
golangci-lint-linter = "golangci_lint_linter:main"

[project.urls]
Repository = "https://github.com/manuelarte/golangci-lint-linter.git"
Issues = "https://github.com/manuelarte/golangci-lint-linter/issues"
Changelog = "https://github.com/manuelarte/golangci-lint-linter/blob/master/CHANGELOG.md"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "_version.py"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.0",
]
